+++ odoo/local-src/connector_esb/controllers/stock.py
@@ -69,3 +69,28 @@ class StockController(http.Controller):
                 .component('ws.message.product.stock.cnk')\
                 .get_message(products)
             return res
+
+    @http.route('/connector_esb/stock/sku',
+                type='json', auth='user', csrf=False)
+    def product_stock_sku(self, products=None):
+        """ Return stock levels of all products (use the SKU)
+        or for specific products
+
+        Expect a POST with multipart/x-www-form-urlencoded
+        The stock levels are returned for all products or for specific
+        products if the form field ``products[]`` is filled with SKU::
+
+            $ curl -X POST \
+                    http://localhost:8069/connector_esb/stock/sku"
+
+        """
+        ensure_db()
+        request.uid = SUPERUSER_ID
+        env = request.env
+
+        backend = env['esb.backend'].sudo().get_singleton()
+        with backend.work_on('product.product') as work:
+            res = work\
+                .component('ws.message.product.stock.sku')\
+                .get_message(products)
+            return res